[ISSUE-257] RssMRUtils#getBlockId change the partitionId of int type to long (#266)
### What changes were proposed in this pull request?
In the RssMRUtils#getBlockId method, change the partitionId of int type to long, make sure no overflow

### Why are the changes needed?
1: In the RssMRUtils#getBlockId method, change the partitionId of int type to long, make sure no overflow


### Does this PR introduce _any_ user-facing change?
Fix #257 issue ,view this issue for details


### How was this patch tested?
org.apache.hadoop.mapreduce.RssMRUtilsTest#partitionIdConvertBlockTest

+++ client-mr/src/main/java/org/apache/hadoop/mapred/SortWriteBufferManager.java
@@ -312,7 +312,7 @@ public class SortWriteBufferManager<K, V> {
     final byte[] compressed = RssShuffleUtils.compressData(data);
     final long crc32 = ChecksumUtils.getCrc32(compressed);
     compressTime += System.currentTimeMillis() - start;
-    final long blockId = RssMRUtils.getBlockId(partitionId, taskAttemptId, getNextSeqNo(partitionId));
+    final long blockId = RssMRUtils.getBlockId((long)partitionId, taskAttemptId, getNextSeqNo(partitionId));
     uncompressedDataLen += data.length;
     // add memory to indicate bytes which will be sent to shuffle server
     inSendListBytes.addAndGet(wb.getDataLength());
